Fix: initial working implementation of mutex locks
This commits adds a 'pause' instruction in the busywait loop rather than
just keeping it empty in the threads.c file. This makes the locks work,
but the stateful_cr test routine faults with a divide by zero fault after one of the threads finishes printing. Need to triage that bug still, however, we have some sync primitive now.