add support for thread return values
scheduler now supports threads that can return a generic pointer (void
*) that will be saved in the task struct of the exiting thread together
with the exit code. Finalize makes no attempts at verifying this pointer
or the memory it points to. Could lead to leaks if the memory is never
deallocated, but we can do that by tracking when the parent thread
exits, and then reaping all the children threads' resoures. This memory
could also be tied to the same address space as that of the calling
thread once we can vmem support. TODO